tests: Fix a mipsel test failure on test 05-sim-long_jumps
The accept syscall on mipsel is syscall number 168, and multiplexed
socket syscalls aren't allowed with strict rules.  This conflicts with
test 05-sim-long_jumps as it is currently written.

Reduce the range of syscalls in test 05 from 1-200 to 1-167.  This
should still give us adequate long jump coverage while removing the
collision with mipsel socket calls.

This resolves GH issue #347:
* https://github.com/seccomp/libseccomp/issues/347

Signed-off-by: Tom Hromatka <tom.hromatka@oracle.com>